% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{circle_matrix}
\alias{circle_matrix}
\title{Print numbers in circle}
\usage{
circle_matrix(max_row, max_col)
}
\arguments{
\item{max_row}{Maximum rows of the matrix}

\item{max_col}{Maximum columns of the matrix}
}
\value{
A matrix with ascending numbers in circle
}
\description{
This function helps to create a n*m matrix
with ascending numbers printed in circle.
}
\details{
Steven received an interview question:
“how to create a matrix in circle?”
Steven failed to answer this question during the interview,
but he created a package once he got back to home.
Users need to specify the dimension (n*m) of the matrix, and
the function will print out a n*m matrix and automatically
fill in ascending numbers in circle.
}
\examples{
circle_matrix(4,4)
circle_matrix(6,5)
}
